requirements.txt
geopandas
pandas
folium
camelot-py[cv]
numpy
branca